initial value for ext-elementTracker
The ext-element tracker was firing up with the input blank. Better to
put an initial value parameter as well and insert it with jQuery.

Todo: Try and figure out a way to find out the average spacing between
selected elements and inject that into the input - This way the user can
now when selecting elements what amount of tracking they currently
have(before actually tracking). Similar to how path_x, path_y work. I've
tried to do this before by averaging the distance between selected
elements, but injecting that value into the tracking input element also
triggers the callback which starts the tracking.