ext-inputPlusMinus never attach on zoom
Zoom is a drag-input but it must be skipped from appending +/- buttons
as it is really unnecessary and ugly. The user can click on the zoom
tool and click to incrementally zoom - thus no need for +/- buttons. If
need arises I can create a custom class for it though, just like I did
for the circular angle element.